// web/src/lib/search/interfaces.ts
export interface BaseFilters {
  source_type?: ValidSources[] | null;
  document_set?: string[] | null;
  // Bounds are ISO date strings.
  updated_at_range?: { start: string | null; end: string | null } | null;
  tags?: Array<{ tag_key: string; tag_value: string }> | null;
}
// ...
export interface SearchDocWithContent {
  document_id: string;
  chunk_ind: number;
  semantic_identifier: string;
  link: string | null;
  blurb: string;
  source_type: ValidSources;
  boost: number;
  hidden: boolean;
  metadata: Record<string, string | string[]>;
  score: number | null;
  is_relevant?: boolean | null;
  relevance_explanation?: string | null;
  match_highlights: string[];
  updated_at: string | null; // ISO date string
  primary_owners?: string[] | null;
  secondary_owners?: string[] | null;
  is_internet: boolean;
  content?: string | null;
}